Begin refactoring.  Most logic for path comparison is now moved to app.paths.
